+++ facet-trait/CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Other
+
+- Removed `Facet::ARCHETYPE`
+
 ## [0.2.5](https://github.com/facet-rs/facet/compare/facet-trait-v0.2.4...facet-trait-v0.2.5) - 2025-04-10
 
 ### Other